Module plactic_monoid::monico

Re-exports

pub use crate::tab::*;

Functions

D 🔒

fonction D de Monico (3.1). Majorant du nombre de [x] tq d(w,x)=1

dist 🔒

distance entre 2 mots de même taille

dist_float 🔒

distance entre 2 mots quelconques

dist_tab_tab 🔒

distance entre 2 tableaux contenant le même nombre d’éléments

dist_tab_word 🔒

distance entre un tableau et un mot de meme taille

f 🔒

renvoie le cardinal de l’intersection entre le support de w et {1,..,s} avec supp(w) = ensemble des symboles distincts composants w

find_q0

trouve un mot q0 tel qu’il existe une permutation sigma telle que [sigma(q0)][b] = [c]

permute 🔒

applique en place une permutation de l’ensemble des perumations définit par Monico au mot w de taille n

plactic_division

Algorithme 1 de Monico. renvoie un mot q tel que [q][b]=[c]

plactic_division_lifting

Algorithme 2 de Monico. renvoie un mot q tel que [q][b]=[c]